requirements.txt
anyio==4.10.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==13.15
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tornado==6.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
